auth update

+++ shop/internal/biz/user.go
@@ -109,9 +109,9 @@ func (uc *UserUsecase) PassWordLogin(ctx context.Context, req *v1.LoginReq) (*v1
 		} else {
 			if passRsp {
 				claims := auth.CustomClaims{
-					ID:          uint(user.ID),
+					ID:          user.ID,
 					NickName:    user.NickName,
-					AuthorityId: uint(user.Role),
+					AuthorityId: user.Role,
 					StandardClaims: jwt2.StandardClaims{
 						NotBefore: time.Now().Unix(),               // 签名的生效时间
 						ExpiresAt: time.Now().Unix() + 60*60*24*30, // 30天过期
@@ -147,9 +147,9 @@ func (uc *UserUsecase) CreateUser(ctx context.Context, req *v1.RegisterReq) (*v1
 		return nil, err
 	}
 	claims := auth.CustomClaims{
-		ID:          uint(createUser.ID),
+		ID:          createUser.ID,
 		NickName:    createUser.NickName,
-		AuthorityId: uint(createUser.Role),
+		AuthorityId: createUser.Role,
 		StandardClaims: jwt2.StandardClaims{
 			NotBefore: time.Now().Unix(),               // 签名的生效时间
 			ExpiresAt: time.Now().Unix() + 60*60*24*30, // 30天过期
